requirements.txt
openai
python-dotenv 
moviepy
azure.ai.textanalytics
azure.core
azure-cognitiveservices-speech
pillow
opencv-python
streamlit
